add type for step
