+++ server/controller/src/main/java/ai/starwhale/mlops/api/protocol/model/ModelVersionVo.java
@@ -17,7 +17,7 @@
 package ai.starwhale.mlops.api.protocol.model;
 
 import ai.starwhale.mlops.api.protocol.user.UserVo;
-import ai.starwhale.mlops.domain.job.spec.StepSpec;
+import ai.starwhale.mlops.domain.job.spec.step.StepSpec;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.swagger.v3.oas.annotations.media.Schema;
 import java.io.Serializable;
