+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/spec/step/StepSpec.java
@@ -14,8 +14,10 @@
  * limitations under the License.
  */
 
-package ai.starwhale.mlops.domain.job.spec;
+package ai.starwhale.mlops.domain.job.spec.step;
 
+import ai.starwhale.mlops.domain.job.spec.ContainerSpec;
+import ai.starwhale.mlops.domain.job.spec.ParameterSignature;
 import ai.starwhale.mlops.domain.job.spec.svc.ServiceSpec;
 import ai.starwhale.mlops.domain.runtime.RuntimeResource;
 import ai.starwhale.mlops.exception.SwValidationException;
@@ -62,6 +64,8 @@ public class StepSpec {
     @NotNull
     @JsonProperty("show_name")
     private String showName;
+
+    private StepType type;
 
     private Integer concurrency = 1;
