+++ server/controller/src/main/java/ai/starwhale/mlops/domain/job/step/bo/Step.java
@@ -18,7 +18,7 @@ package ai.starwhale.mlops.domain.job.step.bo;
 
 import ai.starwhale.mlops.common.TimeConcern;
 import ai.starwhale.mlops.domain.job.bo.Job;
-import ai.starwhale.mlops.domain.job.spec.StepSpec;
+import ai.starwhale.mlops.domain.job.spec.step.StepSpec;
 import ai.starwhale.mlops.domain.job.step.status.StepStatus;
 import ai.starwhale.mlops.domain.system.resourcepool.bo.ResourcePool;
 import ai.starwhale.mlops.domain.task.bo.Task;
