+++ server/controller/src/main/java/ai/starwhale/mlops/domain/task/bo/TaskRequest.java
@@ -16,7 +16,7 @@
 
 package ai.starwhale.mlops.domain.task.bo;
 
-import ai.starwhale.mlops.domain.job.spec.Env;
+import ai.starwhale.mlops.domain.job.spec.step.Env;
 import ai.starwhale.mlops.domain.runtime.RuntimeResource;
 import java.util.ArrayList;
 import java.util.List;
