+++ server/controller/src/test/java/ai/starwhale/mlops/schedule/impl/contianer/CustomContainerSpecificationTest.java
@@ -20,8 +20,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 import ai.starwhale.mlops.domain.job.spec.ContainerSpec;
-import ai.starwhale.mlops.domain.job.spec.Env;
-import ai.starwhale.mlops.domain.job.spec.StepSpec;
+import ai.starwhale.mlops.domain.job.spec.step.Env;
+import ai.starwhale.mlops.domain.job.spec.step.StepSpec;
 import ai.starwhale.mlops.domain.job.step.bo.Step;
 import ai.starwhale.mlops.domain.task.bo.Task;
 import ai.starwhale.mlops.schedule.impl.container.ContainerCommand;
